feat: add control for skippiong
