+++ apps/dashboard/app/new-2/constants.ts
@@ -1,3 +1,7 @@
+export const WORKSPACE_ID_PARAM = "workspaceId";
+export const KEY_PARAM = "key";
+export const API_ID_PARAM = "apiId";
+
 export type StepInfo = {
   title: string;
   description: string;
